requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
ipython==7.18.1
django-filter~=2.4.0